requirements.txt
agentops
openai-agents
python-dotenv
pandas
fastapi
uvicorn[standard]